bổ sung thuộc tính có thể null

+++ RoomBooking.BLL/Services/BookingRoomService.cs
@@ -435,7 +435,7 @@ namespace RoomBooking.BLL.Services
                 string[] periods = item.SlotTime.Split(',');
                 if (periods.Length < 4)
                 {
-                    list.AddRange(ConverDataToTimeSlot(item, periods, item.Times));
+                    list.AddRange(ConverDataToTimeSlot(item, periods, (int)item.Times));
                 }
 
             }
@@ -463,7 +463,7 @@ namespace RoomBooking.BLL.Services
                 int ca = 0;
                 if (item.Times == 5)
                 {
-                    ca = item.Times;
+                    ca = (int)item.Times;
 
                 }
                 else
@@ -502,7 +502,7 @@ namespace RoomBooking.BLL.Services
             }
             else
             {
-                int ca = item.Times;
+                int ca = (int)item.Times;
                 int j = ca - 1;
                 for (int i = ca - 1; i <= item.Times; i++)
                 {
